mahalanobis distance refactor
